Added project structure
Committed the main files for the projects, including the main script, the loading, processing and validation scripts, as well as the tests for these scripts. Also added the pyproject.toml file for the project configuration.

More tests will follow at a later stage, once the main functionality is implemented. The main script is currently empty, but will be filled with the main functionality of the project.

The data is now available via zenodo for testing the import pipeline.